+++ lib/mino/tasks/builtin.clj
@@ -430,7 +430,9 @@
                    ["closure.c"          "stencil_op_closure"]
                    ["push_env.c"         "stencil_op_push_env"]
                    ["pop_env.c"          "stencil_op_pop_env"]
-                   ["env_bind.c"         "stencil_op_env_bind"]]
+                   ["env_bind.c"         "stencil_op_env_bind"]
+                   ["nth_vec.c"          "stencil_op_nth_vec"]
+                   ["first_vec.c"        "stencil_op_first_vec"]]
         stencil-dir "src/eval/bc/stencils"
         disk-files  (->> (file-seq stencil-dir)
                          (filterv #(str/ends-with? % ".c"))
@@ -637,7 +639,9 @@
                    ["closure.c"          "stencil_op_closure"]
                    ["push_env.c"         "stencil_op_push_env"]
                    ["pop_env.c"          "stencil_op_pop_env"]
-                   ["env_bind.c"         "stencil_op_env_bind"]]]
+                   ["env_bind.c"         "stencil_op_env_bind"]
+                   ["nth_vec.c"          "stencil_op_nth_vec"]
+                   ["first_vec.c"        "stencil_op_first_vec"]]]
     (sh! "mkdir" "-p" gen-dir)
     (sh! "mkdir" "-p" tmpdir)
     ;; First stencil writes the preamble; subsequent ones append onto