+++ src/eval/bc/jit/internal.h
@@ -196,6 +196,10 @@ mino_val_t **mino_jit_call_known_prim_slow(mino_state_t *S,
                                            unsigned arg_base,
                                            unsigned argc,
                                            unsigned dst);
+mino_val_t **mino_jit_nth_vec_slow(mino_state_t *S, mino_val_t **regs,
+                                   unsigned a, unsigned b, unsigned c);
+mino_val_t **mino_jit_first_vec_slow(mino_state_t *S, mino_val_t **regs,
+                                     unsigned a, unsigned b);
 mino_val_t  *mino_jit_tailcall_slow(mino_state_t *S, mino_val_t **regs,
                                     unsigned fn_reg, unsigned argc);
 mino_val_t **mino_jit_closure_slow(mino_state_t *S, mino_val_t **regs,